+++ packages/slice-layout/src/templates/verticallayout/index.js
@@ -1,17 +1,17 @@
 import React, { Fragment } from "react";
-import { View } from "react-native";
+import { TcView } from "@times-components/utils";
 import PropTypes from "prop-types";
 import { ItemRowSeparator } from "../shared";
 
 const VerticalLayout = ({ style, tiles }) => (
-  <View style={style}>
+  <TcView style={style}>
     {tiles.map((tile, index) => (
       <Fragment key={`${tile.props.tileName}`}>
         {tile}
         {index !== tiles.length - 1 ? <ItemRowSeparator /> : null}
       </Fragment>
     ))}
-  </View>
+  </TcView>
 );
 
 VerticalLayout.propTypes = {