+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.10.8](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.10.7...@times-components/article-list@9.10.8) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.10.7](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.10.6...@times-components/article-list@9.10.7) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.10.6](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.10.5...@times-components/article-list@9.10.6) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.10.5](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.10.4...@times-components/article-list@9.10.5) (2022-06-21)
 
 **Note:** Version bump only for package @times-components/article-list