+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.49.24](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.49.23...@times-components/article-main-comment@2.49.24) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.49.23](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.49.22...@times-components/article-main-comment@2.49.23) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.49.22](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.49.21...@times-components/article-main-comment@2.49.22) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.49.21](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.49.20...@times-components/article-main-comment@2.49.21) (2022-06-21)
 
 **Note:** Version bump only for package @times-components/article-main-comment