+++ packages/article/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.8.155](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.154...@times-components/article@7.8.155) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.8.154](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.153...@times-components/article@7.8.154) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.8.153](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.152...@times-components/article@7.8.153) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.8.152](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.151...@times-components/article@7.8.152) (2022-06-21)
 
 **Note:** Version bump only for package @times-components/article