+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.7.13](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.7.12...@times-components/author-profile@6.7.13) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
+## [6.7.12](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.7.11...@times-components/author-profile@6.7.12) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
+## [6.7.11](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.7.10...@times-components/author-profile@6.7.11) (2022-06-21)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 ## [6.7.10](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.7.9...@times-components/author-profile@6.7.10) (2022-06-21)
 
 **Note:** Version bump only for package @times-components/author-profile